refactor(core): Resolve circular dependency causing TypeError
Refactored the addon's initialization sequence to break a circular dependency between `playerDataManager`, `placeholderManager`, and `floatingTextManager`.

The `TypeError: not a function` error in `floatingTextManager` was caused by a race condition during module loading. Both `playerDataManager` and `floatingTextManager` were importing `placeholderManager`, but `playerDataManager` was also being imported by `main.js` before the addon was fully initialized. This created a dependency loop that resulted in partially initialized modules and corrupted the `system` object from the `@minecraft/server` API.

This commit resolves the issue by centralizing the placeholder registration logic within the main addon entry point (`main.js`). This enforces a clean, linear initialization order, ensuring all modules are fully loaded before their functions are called.

- Moved the `registerPlayerDataPlaceholders` function logic from `playerDataManager.js` to `main.js`.
- Updated `main.js` to import dependencies directly and call the registration function within `initializeManagers`.
- Removed the now-unused imports from `playerDataManager.js` and cleaned up the linter warnings.
