+++ AddonExeBP/scripts/core/main.js
@@ -17,7 +17,9 @@ import { initializeSpawnProtection } from '../modules/detections/spawnProtection
 import { initializeXrayDetection } from '../modules/detections/xrayDetection.js';
 import { restartAnnouncer } from '../modules/commands/announcement.js';
 import { floatingTextManager } from './floatingTextManager.js';
-import { registerPlayerDataPlaceholders } from './playerDataManager.js';
+import { getLeaderboard } from './playerDataManager.js';
+import { registerPlaceholder } from './placeholderManager.js';
+import { formatCurrency } from './utils.js';
 import '../modules/commands/index.js';
 import './mobDeathEvents.js';
 
@@ -86,7 +88,30 @@ function initializeManagers() {
     infoLog('[AddonExe] Initializing managers...');
     rankManager.initialize();
     initializePunishmentManager();
-    registerPlayerDataPlaceholders(); // Must be before managers that use placeholders
+
+    // Register placeholders here to break the circular dependency between playerDataManager and placeholderManager
+    registerPlaceholder('topbal', ({ index, valueKey }) => {
+        const config = getConfig();
+        const leaderboard = getLeaderboard();
+        const baltopLimit = config.economy.baltopLimit ?? 10;
+
+        if (valueKey === 'list') {
+            const listLimit = Math.min(5, baltopLimit);
+            const topPlayers = leaderboard.slice(0, listLimit);
+            if (topPlayers.length === 0) {return '§cNo players on the leaderboard yet.§r';}
+            return topPlayers.map((player, i) => `§e#${i + 1} §r${player.name}: §a${formatCurrency(player.balance)}§r`).join('\n');
+        }
+
+        if (index >= 0) {
+            if (index >= baltopLimit || index >= leaderboard.length) {return 'N/A';}
+            const playerData = leaderboard[index];
+            if (!playerData) {return 'N/A';}
+            if (valueKey === 'name') {return playerData.name;}
+            if (valueKey === 'value') {return `§a${formatCurrency(playerData.balance)}§r`;}
+        }
+        return 'N/A';
+    });
+
     floatingTextManager.initialize();
     // Clear any expired data on startup
     clearExpiredPunishments();
